+++ src/cli/config.rs
@@ -2,4 +2,5 @@ use crate::api;
 
 pub struct Config<'a> {
     pub api: &'a api::Client<'a>,
+    pub diagnostic_path: Option<String>,
 }